Fix spacing of highlight notes - when populated and when they don't exist
