+++ src/assets/defaultHighlightTemplate.njk
@@ -1,4 +1,7 @@
+{% trim %}
 {{text}} — location: [{{location}}]({{appLink}})
-{%- if note %}{{note}}{% endif %}
+
+{% if note %}{{note}}{% endif %}
+{% endtrim %}
 
 ---
