+++ src/renderer/nunjucks.extensions.ts
@@ -1,0 +1,53 @@
+import nunjucks from 'nunjucks';
+
+function TrimEmptyLinesExtension(): void {
+  this.tags = ['trimEmptyLines'];
+
+  this.parse = function (parser, nodes) {
+    const tok = parser.nextToken(); // Get the tag token
+
+    // Parse the args and move after the block end.
+    const args = parser.parseSignature(null, true);
+    parser.advanceAfterBlockEnd(tok.value);
+
+    // Parse the body
+    const body = parser.parseUntilBlocks('trimEmptyLines', 'endtrimEmptyLines');
+    parser.advanceAfterBlockEnd();
+
+    // Actually do work on block body and arguments
+    return new nodes.CallExtension(this, 'run', args, [body]);
+  };
+
+  this.run = function (_context, bodyCallback) {
+    const rawCode: string = bodyCallback();
+    const rawCodeNoLines = rawCode.replace(/(^[ \t]*\n)/gm, '').trim();
+    return new nunjucks.runtime.SafeString(rawCodeNoLines);
+  };
+}
+
+function TrimLinesExtension(): void {
+  this.tags = ['trim'];
+
+  this.parse = function (parser, nodes) {
+    const tok = parser.nextToken(); // Get the tag token
+
+    // Parse the args and move after the block end.
+    const args = parser.parseSignature(null, true);
+    parser.advanceAfterBlockEnd(tok.value);
+
+    // Parse the body
+    const body = parser.parseUntilBlocks('trim', 'endtrim');
+    parser.advanceAfterBlockEnd();
+
+    // Actually do work on block body and arguments
+    return new nodes.CallExtension(this, 'run', args, [body]);
+  };
+
+  this.run = function (_context, bodyCallback) {
+    const rawCode: string = bodyCallback();
+    const rawCodeNoLines = rawCode.trim();
+    return new nunjucks.runtime.SafeString(rawCodeNoLines);
+  };
+}
+
+export { TrimEmptyLinesExtension, TrimLinesExtension };
